store/requirements.txt
fastapi
uvicorn
redis-om
requests
pydantic-settings
python-dotenv